refactor: Move rule.Rule to public linter package
