+++ _example/plugin/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"github.com/yoheimuta/protolint/_example/plugin/customrules"
 	"github.com/yoheimuta/protolint/internal/addon/rules"
-	"github.com/yoheimuta/protolint/internal/linter/rule"
+	"github.com/yoheimuta/protolint/linter/rule"
 	"github.com/yoheimuta/protolint/plugin"
 )
