+++ internal/linter/linter.go
@@ -2,8 +2,8 @@ package linter
 
 import (
 	"github.com/yoheimuta/go-protoparser/parser"
-	"github.com/yoheimuta/protolint/internal/linter/rule"
 	"github.com/yoheimuta/protolint/linter/report"
+	"github.com/yoheimuta/protolint/linter/rule"
 )
 
 // Linter represents the protocol buffer linter with some rules.
