+++ linter/rule/rule.go
