+++ plugin/grpcServer.go
@@ -2,7 +2,7 @@ package plugin
 
 import (
 	"github.com/yoheimuta/protolint/internal/addon/plugin/shared"
-	"github.com/yoheimuta/protolint/internal/linter/rule"
+	"github.com/yoheimuta/protolint/linter/rule"
 
 	"github.com/hashicorp/go-plugin"
 )
